+++ packages/rotif-nest/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @devgrid/rotif-nest
 
+## 0.1.6
+
+### Patch Changes
+
+- fix discovery service
+
+## 0.1.5
+
+### Patch Changes
+
+- fix module initialization
+
+## 0.1.4
+
+### Patch Changes
+
+- fix discovery service
+
+## 0.1.3
+
+### Patch Changes
+
+- remove terminus
+
+## 0.1.2
+
+### Patch Changes
+
+- add import of terminus module
+
 ## 0.1.1
 
 ### Patch Changes